Huge commit several fixes: App Class size Mini map thread Timers Shader path Wind Mill qSort

+++ CoreEngine_Tutorials/DXENGINE_027/core/src/DX11/DX11Class.cpp
@@ -466,15 +466,15 @@ bool DirectX::DX11Class::OnInit(int g_USE_MONITOR, /*HWND*/void* hwnd, int scree
 	//Init Step: 3, 4
 	ASSERT(createDevice());
     
-	//Init Step: 5 - Get Best Shader of this Graphic Card: dx10,dx10.1,dx11,etc... OUTPUT: ShaderModel
+	//Init Step: 5 - Get Best SHADER of this Graphic Card: dx10,dx10.1,dx11,etc... OUTPUT: ShaderModel
 	getProfile(g_USE_MONITOR);
 
 	//Init Step: 6 Before Resize (SetCamera2D & SetCamera3D)!
 		Initialize3DCamera();
 
 	//Init Step: 7 - CreateWindowSizeDependentResources
-	// Creates a render target view and depth stencil surface/view per swapchain
-	for (size_t m = 0; m < SystemHandle->windowsArray.size(); m++)
+	// Creates a render target view and depth stencil surface/view per SWAPCHAIN
+	for (UINT m = 0; m < SystemHandle->windowsArray.size(); m++)
 	{
 		ASSERT(Resize(m, screenWidth, screenHeight, screenNear, screenDepth, fullscreen, depthBits));
 	}
@@ -491,8 +491,8 @@ bool DirectX::DX11Class::OnInit(int g_USE_MONITOR, /*HWND*/void* hwnd, int scree
 
   #if defined INTRO_DEMO || defined USE_ALPHA_BLENDING // 26
 	//Init Step: 13 Create and Set the depth stencil state: With the created depth stencil state we can now set it so that it takes effect:
-	ASSERT( createSetDepthStencilState (true) );	// Zbuffer on state
-	ASSERT( createSetDepthStencilState (false) );	// Zbuffer off state
+	ASSERT( createSetDepthStencilState (true) );	// ZBUFFER on state
+	ASSERT( createSetDepthStencilState (false) );	// ZBUFFER off state
 
 	//Init Step: 14 - Transparency: To render text on top of 3D
 	ASSERT(CreateBlendState());
